pyproject.toml
[project]
name = "volleyball-play-no-play-model"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.44.0",
    "av>=11.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pillow>=12.0.0",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "onnxscript>=0.5.6",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "ultralytics>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]
